Ran some tests on CustomerService functionality and implemented corrections to prevent errors. Now it is all working!

+++ Business/Services/CustomerService.cs
@@ -1,6 +1,7 @@
 ﻿using Business.Factories;
 using Business.Interfaces;
 using Business.Models;
+using Data.Entities;
 using Data.Interfaces;
 
 namespace Business.Services;
@@ -9,12 +10,14 @@ public class CustomerService(
     ICustomerRepository customerRepository,
     IContactInfoRepository contactInfoRepository,
     IAddressRepository addressRepository,
-    ICustomerAddressRepository customerAddressRepository) : ICustomerService
+    ICustomerAddressRepository customerAddressRepository,
+    IProjectRepository projectRepository) : ICustomerService
 {
     private readonly ICustomerRepository _customerRepository = customerRepository;
     private readonly IContactInfoRepository _contactInfoRepository = contactInfoRepository;
     private readonly IAddressRepository _addressRepository = addressRepository;
     private readonly ICustomerAddressRepository _customerAddressRepository = customerAddressRepository;
+    private readonly IProjectRepository _projectRepository = projectRepository;
 
     public async Task<bool> CreateCustomerAsync(CustomerRegistrationModel model)
     {
@@ -25,7 +28,11 @@ public class CustomerService(
 
         try {
         await _customerRepository.CreateAsync(customer);
-        await _contactInfoRepository.CreateAsync(contactInfo);
+
+            if (customer != null)
+                contactInfo.CustomerId = customer.Id;
+
+            await _contactInfoRepository.CreateAsync(contactInfo);
         foreach (var address in addresses)
             await _addressRepository.CreateAsync(address);
         foreach (var customerAddress in customerAddresses)
@@ -64,13 +71,17 @@ public class CustomerService(
             if (customer == null)
                 return false;
 
-            var contactInfo = await _contactInfoRepository.GetAsync(ci => ci.Id == customer.ContactInfo.Id);
-            var addresses = await _addressRepository.GetAllAsync(a => a.CustomerAddresses.Any(ca => ca.CustomerId == id));
+            var contactInfo = customer.ContactInfo is not null
+                ? await _contactInfoRepository.GetAsync(ci => ci.Id == customer.ContactInfo.Id)
+                : null;
+
+            var addresses = await _addressRepository.GetAllAsync(a => a.CustomerAddresses.Any(ca => ca.CustomerId == id))
+                ?? new List<AddressEntity>();
 
             CustomerFactory.UpdateEntities(customer, updatedModel, contactInfo!, addresses);
 
             await _customerRepository.UpdateAsync(customer);
-            if (contactInfo != null)
+            if (contactInfo is not null)
                 await _contactInfoRepository.UpdateAsync(contactInfo);
             foreach (var address in addresses)
                 await _addressRepository.UpdateAsync(address);
@@ -82,10 +93,20 @@ public class CustomerService(
             throw new Exception("Could not update customer", ex);
         }
     }
+
+
     public async Task<bool> DeleteCustomerAsync(int id)
     {
         try
         {
+            // Checks if customer has related projects
+            var hasProjects = await _projectRepository.GetAsync(p => p.Customer.Id == id) is not null;
+            if (hasProjects)
+            {
+                Console.WriteLine($"Cannot delete customer {id} because they have related projects. Please first delete any projects related to this customer and then try again.");
+                return false;
+            }
+
             return await _customerRepository.DeleteAsync(c => c.Id == id);
         }
         catch (Exception ex)
@@ -93,4 +114,5 @@ public class CustomerService(
             throw new Exception("Could not delete customer", ex);
         }
     }
+
 }
